avances en solicitudes especificas

+++ src/main/java/com/andyg/bazar/persistence/model/ProductoEntity.java
@@ -20,8 +20,7 @@ public class ProductoEntity {
 
     private String nombre;
     private String marca;
-    private double precio;
+    private int precio;
     private int stock;
-
 
 }
